requirements.txt
requests
python-telegram-bot
elasticsearch
logstash
psycopg2-binary
python-dotenv
